Initial commit of new Jenkins app
This App is based on the blog https://blog.openshift.com/continuous-integration-deployment-v3/, this will deploy to the Library and run a replication controller and service.
The test ruby app is included in the Kubernetes.json but the steps in the guide had an issue updating the Jenkins config so requires more work.